Fix bug when deleted lines are close by ...
When 2 lines are removed with a 1 line gap between them (say 17 and 19) then the line between them (18) will try to be occupied by both the deleted_top and deleted_bottom icon. Only 1 icon is able to fill this position tho and looks ugly. So I added a new deleted_dual icon that is both deleted icons combined to use in this situation.