Fix git compare defaulting to empty string
When grabbing the setting, we are setting the default to HEAD, however
the value from default settings is an empty string. Even though an empty
string is a falsy value, sublime settings will still use it and not the
default. So set the default to HEAD. Also in the get_compare method, do
another check for falsy values, this way if the user manually sets the
value to something falsy it still gets set to HEAD.