Fix for issue #183
Write the UTF-8 BOM to the diff buffer file if the view is "UTF-8 with
BOM".
Should be safe, as the "UTF-8 with BOM" encoding cannot be set/changed
from the GUI.
ST3 does not include the BOM in self.view.

Adding a BOM to the beginning of a file will still show the first line
as modified.