Add GitGutterSettings class that encapsulates settings (#230 related)
This is first step towards rewrite of the code to have less globals and more
passing of objects around.

Created GitGutterSettings class which is created in ViewCollection and passed
onto GitGutterHandler's created from there.

The goal is to remove the ViewCollection and pass settings directly from the
base *Command class. Currently settings object is still created in a couple of
places (ViewCollection, individual "WindowCommand"s and EventListener). In the
end it will be created only from the base TextCommand and EventListener.