Revert back to using separate commands for each actions
We still need to redirect to main GitGutterCommand as only that one creates
an instance of GitGutterHandler.

The advantage is full backwards compatibility and the fact that individual
commands can be potentially disabled based on some condition (currently they
are not).

The only difference to original code is that these are TextCommands and not
WindowCommands so that they automatically work in context of current view.