Fixup git_binary_path_fallback handling
During #315 I've introduced a small issue with handling
`git_binary_path_fallback`. The static
`GitGutterSettings.git_binary_path_fallback` was assigned but never read (we
read from `self._git_binary_path_fallback` instead which was always None). The
only issue that stemmed from it, was that we recalculated git path every time
the settings were changed (so not that big of a deal) rather than only once.

Fixed by using same property when both setting and getting it. Also made the
property instance-local as there is only one instance of Settings currently so
there is no need to use static properties. Same goes for
`git_binary_path_error_shown`.

(The code, as before, won't update git_binary_path when user changes the config
file but that is a separate issue.)