Determine, wheter a view really exists on disk.
Thich change fixes an issue where all lines of a new file are marked as `added`.

## Issue descriptions

The edit_settings(..) command of Sublime Text or the `Side-By-Side Settings` plugin as an example
can create new views for user settings which have a valid file_name attached but might not yet be saved to disk.
If the file is created within a valid git repo, it is neither recognized as ignored nor as untracked.
This results in all lines being marked as added, which is wrong as its state is not clear.