Create only one instance of GitGutterSettings (fixes #334)
It's important that all code shares the same instance of GitGutterSettings.
One reason that this is needed is when we trigger an error on not finding
a git binary. With multiple instances, all will check their own state
of _git_binary_path_error_shown property and we'll end up with multiple
message dialogs.

Fix by checking globals and avoiding creation of GitGutterSettings instance
if it already exists. This is needed as importing git_gutter_settings.py
always re-evaluates the code and that ended up creating new instances.