Securing git execution.
The documentation about python's `subprocess` module makes some recommendations about how to handle subprocess calls, which are applied by this commit.

### Python says:

#### 1. communicate()
> Use communicate() rather than .stdin.write, .stdout.read or .stderr.read to avoid deadlocks due to any of the other OS pipe buffers filling up and blocking the child process.
So maybe we should use it.

#### 2. handle timeouts
>The child process is not killed if the timeout expires, so in order to cleanup properly a well-behaved application should kill the child process and finish communication.
Added a timeout of 30s.

#### 3. handle OSError
> The most common exception raised is OSError. This occurs, for example, when trying to execute a non-existent file. Applications should prepare for OSError exceptions.
So print the exception text to console and handle as empty output.

To avoid memory leaks the promise will finally always be resolved even if unhandled exceptions occure.