Decode git output directly after reading from standard output.
All git calls return text which needs decoding. The default encoding of git is utf-8.
Only in few cases another codec may be required.

To simplify the process of parsing git's output decoding is moved to `git_gutter_handler.run_command()`,
which got a new parameter `encoding` that defaults to true, meaning text output is the default.

Decoding is understood as part of the called git process and doesn't need to
be handled by the promise chain.