Show the correct branch or tag name in the 'compare against' status message.
The issue:

No matter what branch is selected to compare against the status bar always shows HEAD or the first 7 characters of a commit's hash.

The solution:

1. Save the 'refs/...' string to settings rather than a specific commit.
2. As a result `git diff` is always run against the HEAD of the branch instead of a certain commit.
   If a new commit is checked in, diff automatically updates, as it does for 'Compare against HEAD'.
3. The status bar shows the branch or tag name instead of the commit hash.

To handle it efficiently the methods `get_compare_against()` and `set_compare_against()` are added to git_handler.

By adding those two methods to git_handler and moving all required steps to apply a new compare target over there,
the cross references can be reduced and the git_gutter_compare module becomes far simpler
as well as the calls of its objects.

In `live_mode` or `focus_change_mode` the view is compared against the new branch/commit/tag
by `on_activate` event when returning from a quick panel, so `git_gutter` command doesn't need
to be called then. It is only invoked, if `Compare against HEAD` or `Compare against origin` is called
as they do not show a panel.