Reduce module dependencies between git_gutter.py and its submodules.
1. Remove unnecessary calls of on_disk() in all git_gutter_compare objects:

As all methods in git_gutter_compare are called by the `git_gutter` command,
which already checks if the file is `on_disk()`, it is not required to check this here again.

2. Remove explicit transfers of `view` to git_gutter_compare objects:

As each git_handler object contains a reference to the attached view, it is not necessary
to copy it over to each function again.

3. Import only TextCommand from sublime_plugin.