Fix Issue #189: Disable GitGutter for binary files and scratch views.
As SublimeText's hexadecimal encoded representation of a binary file wouldn't
be saved to disk in a form git diff could handle it and there is no reason to show
markers for those files, the whole git_gutter command is disabled if a view
returns 'Hexadecimal' encoding.

Views marked as scratch are most likely not intended to represent normal files
so handle them like widgets.