Fix Issue #264 "Broken Menu File"
Even though all menu items appear to be ok it might be a clean solution to
add the caption to all existing items, too. This at least avoids some issues
and is well practice in most plugins anyway.

For consistency with existing commands from the Default Package the menu
entries of the Goto Main Menu are renamed to "Goto ... Git Difference"