Internal: Decode git output directly after reading from standard output.
All git calls return text which needs decoding. The default (fall back) encoding
of git is utf-8. Files might be encoded with default code page of the OS and keep
handled as usual.

The main reason for this commit is to simplify the process of parsing git's
output and avoid struggling with decoding again and again in all the places a
git call is handled. This should result in cleaner code and decoding errors
could be catched on one central point within git_handler.

Decoding is understood as part of the called git process and doesn't need to
be handled by the promise chain like in `git_gutter_show_diff`.