Kill git after timeout in ST3 and catch OSErrors only instead of all Exceptions.
If python 3 interpreter is recognized in the import block, the timeout handling
of `Popen` is enabled. (Thanks rchl). Otherwise a dummy TimeoutExpired class is
created for python 2.6

OS errors are caught as documentation wants, but anything else must be prevented
with other solutions. To avoid ValueErrors and IndexErrors due to invalid
`args` elements, the `git_gutter` command is disabled at all, if no
`git_binary_path` was found by settings module.