Fix Issue #189: Disable GitGutter for binary files and scratch views.
There are several situations running GitGutter might be useless or disturbing,
so it should keep disabled.

1. BINARY FILES

As SublimeText's hexadecimal encoded representation of a binary file wouldn't
be saved to disk in a form git diff could handle it and there is no reason to
show markers for those files, the whole git_gutter command is disabled if a
view returns 'Hexadecimal' encoding.

2. SCRATCH / READONLY / WIDGET VIEWS

Views marked as scratch or readonly are most likely not intended to represent
normal files so handle them like widgets and disable GitGutter.

3. UNATTACHED VIEWS

SublimeLinter disables linting for views with no valid window() attached and
states it as possible way to detect widgets for sure in some builts of
SublimeText.

4. SublimeREPL

The sublimeREPL plugin sets a view setting repl to mark a view as output.
SublimeLinter uses this setting to disable itself, so GitGutter should do so,
too. Where to check the view?

The git_handler is now created for all GitGutterCommand objects to
be used in its is_enabled method. This doesn't hurt as temporary
files are created on demand by git_handler.