Feature: Add Side-By-Side settings and support
This commit introduces `git_gutter_open_file` command as wrapper for the
legacy `open_file` command, which is used for backwards compatibility.

Beginning with Sublime Text 3124 the `open_file` commands are hidden as they
are replaced by a more convenient `edit_settings` (Side-By-Side settings).

All available settings are added to the command pallet as
- Preferences: GitGutter Settings
- Preferences: GitGutter Key Bindings
- Preferences: GitGutter Popup Stylesheet

The Main.sublime-menu was updated
- to show Side-By-Side settings
- to order JSON keys the same way for all items.
- use ${platform} as part of the "file" argument to shorten the file
  which is possible as we do only use platform depended key bindings.