Feature: Add support for custom icon packages (Issue #256)
A new setting `theme` provides a value of `<Name>.gitgutter-theme` which is an
empty file passed to `sublime.find_resources(...)` to find the folder containing
the gutter icons to show for the events and the gitgutter_popup.css to use for
the diff popup. So theme developers can add custom icons and popup styles for
GitGutter to their package. The handling is derived from Sublime Text's default
`theme` handling to keep the setting as easy as possible, so user doesn't need
to care about paths.

As Sublime Text 2 does not support for `sublime.find_resources(...)` only
built-in themes are allowed then.

The commit contains two built-in themes `Default` and `Default HiDPI` to support
high resolution displays for older versions of Sublime Text.

Most themes already provide name@2x.png files to support high resolution
displays without any special setting being required. Sublime Text is expected to
use those files in case of DPI scale > 1. So GitGutter may or may not provide the
larger icons with this naming in the `Default` theme as well in case Sublime Text
will support them in the future.