Enhancement: Event driven tracking of changed compare target replaces polling.
Instead of polling the commit hash each 5th or 10th second to find out if the
compare target was changed by an external application the forwarded event_types
from EventListener are used to decide when to.

The basic idea is to invoke git commands and check for changes only if the file
was opened, saved or a view gets activated. This way the number of `git` calls
is reduced and even avoided at all while the user is typing. This is the most
unlike situation for compare targets to change.

A special situation is when a unmodified file is changed while SublimeText is
not active. GitGutter receives `modified`, `activated` and `modified` again
short after each other then, as soon as SublimeText receives focus. Due to
`view.is_loading() = True`, view.is_scratch() = True` and debouncing feature,
the `activated` event is dropped. As GitGutterCommand sees `modified` only, it
would think user is typing and not update cached values. Therefore
`deactivated` is handled to ensure cache is cleared as soon as SublimeText
looses focus.

An alternitive solution would be to use debouncing for `modified` only, as it
is the only command, really worth to do it for. So GitGutterCommand would see
`activated` and could be able to decide whether to run or/and invalidate
cached status information. As this might result in larger changes, it is not
intended to be done in near future.