Fix: Problem with smudge/clean filters (Issue #74)
As discussed in issue #74 git supports smudge filters to manipulate the content
of files when being archived, checked out or added/committed to the repository.
Git applies those filters to a selected set of commands only. Therefore this
commit replaces `git show` by `git archive`. It uses `--format=zip` instead of
the default `--format=tar` as extracting a zip archive works properly with the
ByteIO class.