Fix: Avoid exception due to missing dependencies. (Issue #366)
If Sublime Text starts the very first time with GitGutter installed some
dependencies may not be satisfied yet, which results in an uncaught exception.

To avoid issues with the whole plugin loading process the depending functions
are disabled, if dependencies are not satisfied.

This commit disables the status bar text, if jinja2 library is missing.