README: Refactoring
As the GitGutter-Edge package does no longer exist we need to update the README
to tell people about the new ways to receive pre-releases and nightflies.

We might also add the list of features to the top of GittGutter's description
to give a brief overview about the capabilities.

Due to issue #366 I decided to list and link the required packages in a
"Dependencies" section and give some more information about installation.

All paragraphs are modified to have the same style of writing and usage of
common phrases.