Fix: Noticeable flicker/delay in icons when working with files (Issue #340)
This heavy delay on larger files is mainly caused by slow performance of the
Sublime Text API functions, which are heavily used to determine the regions
to add gutter icons to.

Working directly on the text buffer on an asynchronous task may not be a good
idea at all, as the view might always change during execution.

Therefore the content of the modified part of a view is buffered in a string
which is used to create a map to convert the diff lines to regions without
API calls. Same is done by protected regions.

This commit reduces the usage of API functions as much as possible with the
following results. The benchmark was recorded on Windows 10 x64 with
Sublime Text 3126 using python's cProfile module.

| marked  | total |  1.3.0  |   now |
| ---     | ---   | ---     | ---   |
|       0 |   300 |     3ms |   0ms |
|     150 |   300 |    40ms |   1ms |
|     300 |   300 |    51ms |   2ms |
|       0 |  1300 |     3ms |   0ms |
|     300 |  1300 |    51ms |   2ms |
|    1300 |  1300 |   194ms |   5ms |
|       0 |   65k |     5ms |   0ms |
|       1 |   65k |     5ms |   1ms |
| 1 & 65k |   65k | 8.700ms |  48ms |
| 0...65k |   65k | 8.700ms |  70ms | -> ignored/untracked
| 0...65k |   65k | 8.700ms | 147ms | -> modified