Fix: Resolve symlinks on Windows machines correctly (Issue #373)
This is a fix for the issue of os.path.realpath() not to resolve symlinks on
Windows as it is an alias to abspath() only.

see: http://bugs.python.org/issue9949

Note 1

As git_gutter_handler.py and the GitGutterHandler class grows too much,
the related functions for path handling are moved to a seperate module.

Note 2

With respect to future plans for a modularized plugin structure, the
path functions are put into the folder 'modules'.