Enhancement: Don't call git for diff popup at all.
The git diff result is already cached by updating the gutter icons. Under no
circumstances it is required to explicitly recall diff by hovering the gutter
area. The diff popup can make use of the cached value always. If the diff cache
is empty, no popup is needed as gutter icons are not there, too.