Fix: Ensure not to start more than one evaluation at a time.
Even though preceding optimizations target to minimize latencies of evaluating
file modifications it might happen especially with large files being read from
a slow maybe network file system, that the next run is already triggered,
before the current evaluation has finished.

This causes two or more promise chains to be active at a time with unknown
consequences.

As the structure of the promise chain makes aborting the process hard to
implement, this commit simply skips events as long as the chain is busy.