Fix: Avoid FileNotFoundError exception, if symlink on windows can't be resolved.
Issue:

The function _getfinalpathname added to resolve symlinks on Windows machines
raises an FileNotFoundError exception if the provided path does not exist or
the link target can't be resolved. os.path.realpath on the other hand seems
to resolve targets on Linux even though it does not exist without Exception.

Solution:

Catch the FileNotFoundError and return the provided path as is.

Other:

Always return None if path is None instead of raising an exception.