Fix: Avoid unnecessary git calls when typing.
Due to a regression the `git_gutter_handler._git_temp_file_valid` is never set
true. This has no negative functional effect, but causes the compare target's
commit hash to be read after each modification of the view, which is not
required and a waste of resources and time only.

Furthermore it is not necessary to query the current branch name on normal
user inputs if already valid.

At the moment 3 git calls are performed each time the view is modified:

  git rev-parse HEAD
  git diff ...
  git rev-parse --abbrev-ref HEAD

With this commit the only required one is left:

  git diff ...

This has effect on handling 'modified' event only! In all other situations all
three calls are performed.