Fix: Files are randomly marked as modified (Issue #381)
Issue:

With introducing `git archive` command it was considered save to remove the
eol mangling as correct translations should happen by this command.

This seems not to apply sometimes.

Solution:

Reintroduce eol mangling for the files read from repository.

Note:

As the temporary view file is created from text, it should be written to disk
with `\n` and thus shouldn't need mangling.