Fix: Settings won't apply without restart of ST (Issue #184)
The GitGutterSettings module has 3 big issues:

1. View/Project specific GitGutter settings don't work at all.
2. Settings won't apply without restart of ST (Issue #184).
3. The whole module suffers from ST's double loading bug.

In order to fix these issues, the settings module is completely rewritten.

The basic idea in supporting project based settings is to be able to place each
entry of the GitGutter.sublime-settings file to a Project settings tree as well.

This is achieved by replacing the GitGutterSettings class by a new ViewSettings
class, which is attached to each GitGutterCommand object to handle view
and project specific settings.

To ensure all settings entries are unique (Issue #326) the settings must be
prefixed using 'git_gutter_' when placed in the project settings.

    Example:

        GitGutter.sublime-settings
        {
            show_in_minimap: 3
        }

        Preferences.sublime-settings
        {
            git_gutter_show_in_minimap: 3
        }

        Project.sublime-project
        {
            ...
            settings:
            {
                ...
                git_gutter_show_in_minimap: 3
                ...
            }
        }

This is something which works by nature, if settings are read by the settings.get()
method each time a value is required instead of creating local copies of the values
read once after startup. All settings objects automatically track changes and
provide up to date values. Settings are cached for one evaluation run only.

The settings module is moved to the 'modules' sub directory to avoid it being
loaded twice upon startup.

A global function get() ensures save access to the unique global instance of
the GitGutter.sublime-settings file.