Fix: Print error message and abort evaluation on failed git call.
Again and again people complain about all lines being marked with '+' which is
the result of a failed git call in most cases. The empty output returned in
such situations can't be handled as error and equals to normal situations.

To be able to debug such situations or give users a hint about what's wrong,
'fatal:' error messages returned by git are printed to console. This will show
an error message on each try to read content of untracked files from git at the
moment, but this will be fixed by use of `git status` in the future.

If an exception was raised during git execution the evaluation is aborted by
returning `None` in `decode_diff()` so no gutter icons or status message are
displayed in such cases. If a former evaluation succeeded, this results in no
more updated gutter icons or status message, but this shouldn't happen too
often as git is expected to fail from the first call if some issue occurs here.

The error handling may be improved in the future.