Enhancement: Refactoring EventListener module
Goals:
1. Avoid double loading of modules.
2. Expose minimal set of classes to ST only.
3. Run only one timer at a time for debouncing.
4. Don't loose events due to debouncing.
5. Increase responsiveness.

In order to forward events delayed after some idle time GitGutter called
sublime.set_timeout() on each event until idle time elapsed. This can cause
up to 30 timers to be active at a time, if a user types text very quickly
without a break.

In order to reduce the number of timer invocations the ViewEventListener class
is created, which tracks the state of a unique timer. A timer is started only,
if no one is active yet. All received events are queued until idle time
elapsed and sent to GitGutterCommand then. This way no event gets lost and
Sublime Text API is not stressed unnecessarily. The first debounce timer starts
with a small but fixed delay to increase responsiveness. All following events
are delayed by the `debounce_delay` as long as a timer is active.

The queued events are represented by numbers instead of strings to save some
memory. The event ids are chosen to be combined in a binary mask instead of a
real queue as we only need to know that but not how much events of one type
have been received.

The whole stuff is moved to the 'modules' sub directory to hide as much from
ST as possible. The only exported class is the 'EventListener'. This is done
here for the event listener to respect future changes which are required for
all modules, which are imported by another module to avoid double loading.

What is double loading? ST actively loads all modules located in a package's
root directory to search for API classes to hook. If such a module is imported
by another one (see: 'git_gutter.py') it will be loaded again by the python
interpreter. All classes and globals are initialized again, which is a waste
of time and may cause issues in some cases.