Internal: Refactor 'Compare against ...' module
Goals:

1. Avoid creating temporary objects for simple functions.
2. Clear readable functions which do not rely on overloaded methods.
3. Common interface for each subcommand to create a command map later.

Basically a object orientated approach is desirable, but it seems a bit
overeager to create a temporary object for something which can simply be
a function.

Each GitGutterCompare object is created and initiated passing the git_handler
as argument to __init__(). After that run() is called to perform the real action.
As each object is immediately destroyed after use, this approach seems a bit
misplaced here.

Instead create one function for each sub command with a common interface,
which can be used to assign each command name and function in a dictionary later,
which can act as a command map.