Enhancement: Use git version to check if git works properly (Issue #411)
Execute `git --version` to determine whether to enable GitGutterCommand.

Git version is updated each time the view is activated or a file is loaded/saved.

The git command is run synchronously without Promise() as the version needs
to be available immediately in ´GitGutterCommand.is_enabled()` and there
time to query version should be negligible.

The missing binary is added to a list to show error message only once per
binary. As soon as git is working again, GitGutter removes it from the list
and informs user about it.