Fix: All export-ignored files are marked as new files (Issue #409)
Issue:

Users repetitively report files being marked as new even though they aren't.

Since commit dce471e GitGutter uses `git archive` to read the content of a file
from the tree. This change was made to fix issue #74 and add some basic support
for smudge/clean filters.

But `git archive` doesn't find files which match the `export-ignore` patterns
in the `.gitattributes` or `.git/info/attributes` files. This results in empty
output for a file which is expected to be there and thus will be displayed as
added in the gutter area.

Solution:

The plumping command `git cat-file --filters` reads the file content with
smudge filters applied as well. Thus it is used to replace `git archive` to
fix this issue without breaking smudge/clean filters support.