Fix: Regression with git cat-file and untracked/ignored files.
Commit 4a95320 replaces `git archive` with `git cat-file` to read the content
from a file in git's tree.

In case of `file not found` error `git archive` used to throw an AttributeError
which was used to detect the `untracked` state of a file. `git cat-file` returns
empty file content only as `git show` did in the past.

Therefore the handling of untracked/ignored files is changed to make everything
work properly again.