Other: Remove diff popup border
The border created in CSS will scroll off the display area, if the displayed hunk is large enough and thus can't do what its meant for - contrast.

It's up to the theme to define a global border around popups or not. This can't be handled by CSS. Same applies to borders around the text area.

To ensure contrast the blending colors are tuned a little bit.