Fix: Use git --filters argument only if git 2.11.0+ is available (Issue #416)
Commit 4a95320 introduced `git cat-file --filters` to read content from a tree object and apply smudge filters. The reason for that change was `git archive` ignoring files which are filtered from export in a .gitattributes file.

With release of GitGutter 1.7.0 it turned out that `--filters` argument is supported by git 2.11.0 and above only, which may not be available on some conservative Linux distributions. As a result of the failed execution GitGutter sees an empty file and marks all lines added.

The only solution right now is to use `-p` instead of `--filters` if older git versions are detected. This will reintroduce issue #74 for all users of legacy git setups but there is no other solution available.

Note: smudge filters require git 2.11.0+ beginning with this commit.