Fix: Support Info is unable to read package version and git version
If GitGutter is installed via Package Control as GitGutter.sublime-settings the PACKAGE constant was initialized incorrectly and thus the version number could not be loaded.

The cwd argument of subprocess.POpen is required only to read the commit hash in case a user is running a pulled master branch and thus is omitted if installed as package.

TODO: gitgutter_version() -> sublime.load_resources() not supported by ST2.