Fix: diff popup is not displayed if view is scrolled horizontal.
The diff popup was placed to the first column of the hovered row. If the first column is not visible due to horizontal scroll position, the popup is not created by Sublime Text.

Therefore a new function `_visible_text_point()` is used to return the text point of the first visible column of the hovered row, which is used for popup creation now.