Enhancement: Add support for git diff --histogram algorithm (Issue #422)
Issue #422 proposes the usage of `git diff --histogram` to determine file modifications.

This commit therefore replaces the "patience" setting by a more general "diff_algorithm". This key can be set to ("default", "minimal", "patience", "histogram") to select the desired diff algorithm. The setting is internally translated to the command line argument to ensure only valid arguments are passed to git.