Fix: Reduce gutter icon jumping while editing text
As evaluation of changes and updating the gutter icons is handled asynchronously, the view might change in the meanwhile, rendering the evaluation result invalid or at least inaccurate. As the result lines can't be translated to correct regions anymore. Therefore added regions may jump up or down to the next line.

To reduce this effect two changes are made:

1. initial evaluation delay is increased to 200ms
2. updating regions (gutter icons) is skipped if the view change counter has changed during evaluation