Fix: Gutter icons are not displayed in Sublime Text 2
The issue was introduced with commit 9e5ab88 which intends to reduce gutter icon jumping

The view change counter is used to check whether to update the regions or not. If this counter is not available the gutter icons are never updated as the related function returns true in that case.

This commit changes the behavior of the function to return False if the change counter is not available.