Internal: Log the reason for GitGutter to be disabled
In order to have a chance to determine, why GitGutter keeps quite for certain files, the reason is printed to console, if

  "debug": true

is set. The state itself is handled as integer for performance reasons. The text message is created only, if GitGutter has something to tell. The descriptions are organized using a dict rather than a tuple to provide a clear and obvious association between the state id and the message being printed.