Enhancement: Add command with keybinding to revert changes
With GitGutter's diff popup a user can revert a changed hunk to the state of the compared commit.

This commit moves the code to perform the revert operation into a dedicated module and adds it to the internal command chain:

1. Add the revert.py with the implementation moved from popup.factory module
2. Add "revert_change" to command map in the main GitGutterCommand class.

   provides: `view.run_command("git_gutter", {"action": "revert_change"}`

3. Add TextCommand "git_gutter_revert_change" meant for external use.

   provides: `view.run_command("git_gutter_revert_change"}`

4. Add an entry to the Default.sublime-commands
5. Add Ctrl+Shift+Alt+Z key binding.
6. Add a short note to the README.

NOTE: The revert operation does not work with multiple cursors at the moment.