Fix: Compare against Origin (Issue #371)
Closes: #371

This commit changes the "Compare against Origin" function to store `<branch_name>@{upstream}` as compare target instead of `<branch_name>` which might not be useful in all use cases.

This modification enables a user to compare against the upstream revision of each currently checked out branch.

The command name an the title are not modified for backward compatibility reasons.

If the local branch doesn't have a tracked remote branch, each file is displayed as `inserted`.