Internal: Reduce duplications in command objects
This commit removes duplicate methods `run` which call `self.view.run_command('git_gutter', ...)` to forward commands to the main GitGutterCommand object of a view.

The run() method is moved to the GitGutterBaseCommand class to be implemented once, only. Each command class stores an 'ACTION' class attribute to identify the action to forward to GitGutterCommand.