Enhancement: Add Revert Change to Edit Main Menu
This commit intends to enhance consistency of commands availability. All of them should at least be available via

   - key bindings
   - command pallet
   - main menu (for legacy users)

This commit therefore adds the git_gutter_revert_change command to the main menu with the hopefully clear and meaningful caption "Revert Change to Commit". To keep consistent, the caption is modified in the command pallet and README, too.