Internal: Refactor and sort imports
It is always a mess to rebase or merge branches, which add or remove imports resulting in merge conflicts due to the way multiple imports are arranged.

To solve this issue and make merging work like a charm without conflicts, only one class/module/package is imported per line.

This commit does not provide any functional changes.