Enhancement: Add a command with keybinding to copy content from commit
With GitGutter's diff popup a user can copy the committed content of a hunk to clipboard.

This commit adds the text command "git_gutter_copy_from_commit" to publish this functionality as

   view.run_command("git_gutter_copy_from_commit", {"line": 10})
   view.run_command("git_gutter_copy_from_commit", {"point": 3022})

or:

   view.run_command("git_gutter", {"action": "copy_from_commit", "line": 10})
   view.run_command("git_gutter", {"action": "copy_from_commit", "point": 3022})

It takes an optional argument

  - "line", which is the zero based line number of the hunk to copy or
  - "point", which is the text position of the cursor.

If none of them is provided the line under the first cursor is used.

Also adds:

- main menu entry
- command pallet entry
- key bindings (Ctrl+Shift+c)

NOTE: The operation does not work with multiple cursors at the moment.