Enhancement: Add support for line/point arguments to git_gutter_revert_change
This commit extends the text command "git_gutter_revert_change" to support arguments to explicitly revert changes of a certain line or text point. This may be used by macros or external plugins.

   view.run_command("git_gutter_revert_change", {"line": 10})
   view.run_command("git_gutter_revert_change", {"point": 3022})

or:

   view.run_command("git_gutter", {"action": "revert_change", "line": 10})
   view.run_command("git_gutter", {"action": "revert_change", "point": 3022})

It takes an optional argument

  - "line", which is the zero based line number of the hunk to copy or
  - "point", which is the text position of the cursor.

If none of them is provided the line under the first cursor is used.