Fix: Toggle GitGutter via main menu works with second click only.
The main menu entry to enable/disable GitGutter uses the core command "toggle_setting", which works for settings placed within the "Preferences.sublime-settings" file only. It is used because of the nice check mark it adds to the menu to indicate GitGutter is enabled.

Therefore the toggled setting "git_gutter_enabled" needs to be published to preferences with its default value `true` in order to make the command work properly.

This is achieved by adding a `Preferences.sublime-settings` file with this setting.

As GitGutter always looks for a setting within the Preferences and makes use of GitGutter.sublime-settings only as a fallback, the "enable" entry in `GitGutter.sublime-settings` file will never be queried again and can be removed.

We loose the ability to globally disable GitGutter via "Preferences: GitGutter Settings" command. But as disabling it globally still works via Preferences.sublime-settings and may not make too much sense at all, the step back should not be too big.

Note: The gutter marks may be removed on the next on_activate event, which seems not be sent after menu item click.