Fix: Revert change doesn't work in ST2
The text command "revert_change" was originally designed to work with diff popup only, which is not available for ST2. It uses a command "git_gutter_replace_text" to perform text replacement, which is required as there is no valid edit token available in popups.

The implementation of "git_gutter_replace_text" is moved up to commands.py to make it available in ST2, too.