Fix: Hide "Show Diff Popup" in main menu if not supported
This commit provides required changes to hide the diff popup command from main menu, if the prerequisites for popups are not meat.

The change includes to always publish the `GitGutterDiffPopupCommand` class to Sublime Text and add the `is_visible()` method. As this class was the last remaining in popup commands in order to be created on demand only, it is now moved to `modules/commands.py`. Now all command classes are provided by this one and only module.